worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-tracker-worker"
version = "0.1.0"
description = "A tool to run memory benchmarks on CPython commits"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "click>=8.0.0",
    "memray>=1.0.0",
    "gitpython>=3.1.0",
    "rich>=10.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
memory-tracker = "memory_tracker_worker.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_tracker_worker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 